backend/requirements.txt
# RedMind — Python Dependencies
# Install: pip install -r requirements.txt

# ── Web framework ─────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0

# ── LangChain + LangGraph ─────────────────────────────────────
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0       # OpenAI-compatible (works with OpenRouter)
langgraph>=0.1.0

# ── HTTP client ───────────────────────────────────────────────
httpx>=0.27.0

# ── Redis ─────────────────────────────────────────────────────
redis>=5.0.0

# ── Retry / Circuit Breaker ───────────────────────────────────
tenacity>=8.3.0

# ── Scheduler ─────────────────────────────────────────────────
apscheduler>=3.10.0

# ── Config ───────────────────────────────────────────────────
python-dotenv>=1.0.0